TITANIC SURVIVAL PREDICTION
I have developed a predictive model utilizing the Titanic dataset, which aims to determine the likelihood of a passenger's survival aboard the Titanic. This project is a well-known introductory exercise and benefits from easily accessible data.

The dataset employed in this project encompasses various details pertaining to individual passengers. These details encompass factors such as the passenger's age, gender, ticket class, fare paid, cabin allocation, and crucially, whether the passenger survived the disaster or not.